@@ -1,3 +1,3 @@
 module my-project
 
-go 1.16
+go 1.21.1
