@@ -1,3 +1,3 @@
 module test-example
 
-go 1.16
+go 1.21.1
