@@ -6,8 +6,8 @@ import (
 )
 
 type Person struct {
-	Name string `json: "name"`
-	Age  int    `json: "age"`
+	Name string `json:"name"`
+	Age  int    `json:"age"`
 }
 
 func main() {
